Add support for TOTP over basic auth
Instead of using a static password, allow a TOTP token to be specified
as the password. The interesting side effect of this is that the
resources will be available only for the duration of the TOTP
period. Applications that cache the password for future accesses may
break. However, this also establishes a "natural" timeout as well.